feat: linters
